Extract common functionality in get and post into generic request
...it occurs to me that I don't know the difference between a URL and a URI

UPDATE:

Went and read this http://stackoverflow.com/questions/176264/whats-the-difference-between-a-uri-and-a-url/1984225#1984225
...still have no fucking clue

Found my way to this http://www.ietf.org/rfc/rfc3986.txt
...and decided I didn't actually care that much

Why is it that hard to say "here is the full set of possible
information you can express with these things, and here are the
mandatory and optional ones for a URL and a URI"?

Whatevez.