Make the private methods actually be private
Really, I'm kind of wanting to pull them out into their own classes,
but that would add a lot of files, since the whole thing is like 200 loc,
it doesn't seem to merit that much ceremony. We'll see how I feel about it
as I keep up w/ the lib.